+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [unreleased]
 
+## Added
+- Clean candles instruction related data
+
 ## [v2.20] - 2025-12-17
 
 ## Added